# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LearnSteer Platform REST API",
    "description": "Official Backend API for the **Sasnaka Sansada LearnSteer Platform** \u2014 an educational self-learning, diagnostic examination, and syllabus streaming platform for Sri Lankan A/L students.\n",
    "version": "1.0.0",
    "contact": {
      "name": "Sasnaka Sansada Tech Team",
      "email": "support@learnsteer.lk"
    }
  },
  "servers": [
    {
      "url": "https://ss-quiz-platform-backend-8f9b2c83591e.herokuapp.com/api/v1",
      "description": "Production Heroku Server"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/auth/create-student": {
      "post": {
        "tags": [
          "Authentication & SSO"
        ],
        "summary": "Create a New Student Account (Frictionless Registration)",
        "description": "Collects the 8 essential fields to immediately onboard the student, generating a unique Student ID (e.g. LS-a1b2c3d4) and cartoon-themed Nickname (e.g. BraveLion42).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "first_name",
                  "last_name",
                  "email",
                  "password",
                  "nic",
                  "whatsapp_number",
                  "stream",
                  "medium"
                ],
                "properties": {
                  "first_name": {
                    "type": "string",
                    "example": "Kasun"
                  },
                  "last_name": {
                    "type": "string",
                    "example": "Perera"
                  },
                  "email": {
                    "type": "string",
                    "example": "student@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "SecretPassword123"
                  },
                  "nic": {
                    "type": "string",
                    "example": "200512345678"
                  },
                  "whatsapp_number": {
                    "type": "string",
                    "example": "0771234567"
                  },
                  "stream": {
                    "type": "string",
                    "example": "Bio Science"
                  },
                  "medium": {
                    "type": "string",
                    "example": "Sinhala"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created successfully with Student ID and Nickname"
          },
          "400": {
            "description": "Validation error or duplicate email/NIC/whatsapp"
          },
          "409": {
            "description": "Conflict on unique field"
          }
        }
      }
    },
    "/auth/update-student": {
      "put": {
        "tags": [
          "Authentication & SSO"
        ],
        "summary": "Update Student Profile (Profile Enrichment)",
        "description": "Update secondary demographics once student is registered. Requires JWT Bearer token.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "example": "0712345678"
                  },
                  "district": {
                    "type": "string",
                    "example": "Colombo"
                  },
                  "school": {
                    "type": "string",
                    "example": "Ananda College"
                  },
                  "city": {
                    "type": "string",
                    "example": "Colombo 10"
                  },
                  "al_year": {
                    "type": "string",
                    "example": "2026"
                  },
                  "al_attempt": {
                    "type": "string",
                    "example": "1"
                  },
                  "date_of_birth": {
                    "type": "string",
                    "example": "2007-05-15"
                  },
                  "gender": {
                    "type": "string",
                    "example": "Male"
                  },
                  "guardian_phone": {
                    "type": "string",
                    "example": "0777654321"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication & SSO"
        ],
        "summary": "Student / Admin Login",
        "description": "Authenticates student or admin using Nickname, Email, or NIC with password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifier",
                  "password"
                ],
                "properties": {
                  "identifier": {
                    "type": "string",
                    "example": "BraveLion42"
                  },
                  "password": {
                    "type": "string",
                    "example": "SecretPassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful with JWT token",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsIn...",
                  "user": {
                    "id": 1,
                    "student_id": "LS-9f4a12bc",
                    "nickname": "BraveLion42",
                    "email": "student@example.com",
                    "first_name": "Kasun",
                    "last_name": "Perera",
                    "role": "student",
                    "stream": "Bio Science",
                    "medium": "Sinhala"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/check-nic": {
      "post": {
        "tags": [
          "Authentication & SSO"
        ],
        "summary": "Check If Student NIC Exists",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nic"
                ],
                "properties": {
                  "nic": {
                    "type": "string",
                    "example": "200512345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns exists flag and registration status"
          }
        }
      }
    },
    "/auth/profile/{student_id}": {
      "get": {
        "tags": [
          "Authentication & SSO"
        ],
        "summary": "Fetch Public Demographic Profile by Student ID",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "LS-9f4a12bc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Demographic profile details"
          }
        }
      }
    },
    "/auth/exchange": {
      "post": {
        "tags": [
          "Authentication & SSO"
        ],
        "summary": "Exchange SSO Ticket for JWT Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ticket"
                ],
                "properties": {
                  "ticket": {
                    "type": "string",
                    "example": "sso-ticket-uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Valid JWT token issued"
          },
          "401": {
            "description": "Ticket expired or invalid"
          }
        }
      }
    },
...
    "/b2b/tickets": {
      "post": {
        "tags": [
          "B2B SSO Tickets"
        ],
        "summary": "Issue 60-Second Single-Use SSO Ticket",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nic"
                ],
                "properties": {
                  "nic": {
                    "type": "string",
                    "example": "200512345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SSO ticket issued successfully"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter student JWT token received from /auth/login or /auth/exchange"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Enter B2B API Key configured on the server"
      }
    }
  }
}